Moved Hydra related files into a separate package
