+++ scala/scala-impl/src/org/jetbrains/plugins/hydra/notification/HydraArtifactsNotificationProvider.scala
@@ -1,13 +1,15 @@
-package org.jetbrains.plugins.scala.project.notification
+package org.jetbrains.plugins.hydra.notification
 
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.project.Project
-import com.intellij.ui.{EditorNotificationPanel, EditorNotifications}
 import com.intellij.openapi.util.Key
-import org.jetbrains.plugins.scala.compiler.{CompileServerManager, HydraCompilerSettings}
-import org.jetbrains.plugins.scala.project.notification.HydraArtifactsNotificationProvider._
-import org.jetbrains.plugins.scala.settings.HydraApplicationSettings
+import com.intellij.ui.{EditorNotificationPanel, EditorNotifications}
+import org.jetbrains.plugins.hydra.compiler.HydraCompilerSettings
+import org.jetbrains.plugins.hydra.settings.HydraApplicationSettings
+import org.jetbrains.plugins.scala.compiler.CompileServerManager
 import org.jetbrains.plugins.scala.project.ProjectExt
+import org.jetbrains.plugins.scala.project.notification.AbstractNotificationProvider
+import org.jetbrains.plugins.hydra.notification.HydraArtifactsNotificationProvider._
 
 /**
   * @author Maris Alexandru
@@ -15,7 +17,7 @@ import org.jetbrains.plugins.scala.project.ProjectExt
 class HydraArtifactsNotificationProvider(project: Project, notifications: EditorNotifications)
   extends AbstractNotificationProvider(project, notifications) {
 
-  override def getKey = ProviderKey
+  override def getKey: Key[EditorNotificationPanel] = ProviderKey
 
   override protected def hasDeveloperKit(module: Module): Boolean = {
     val downloadedScalaVersions = HydraApplicationSettings.getInstance().getDownloadedScalaVersions
@@ -32,10 +34,7 @@ class HydraArtifactsNotificationProvider(project: Project, notifications: Editor
     }
   }
 
-  override protected def createTask(module: Module) = new Runnable {
-    override def run(): Unit =
-      CompileServerManager.showHydraCompileSettingsDialog(project)
-  }
+  override protected def createTask(module: Module): Runnable = () => CompileServerManager.showHydraCompileSettingsDialog(project)
 
   override protected def panelText: String = s"No $developerKitTitle artifacts in module. Please download the artifacts."
