+++ scala/scala-impl/src/org/jetbrains/plugins/scala/project/template/Downloader.scala
@@ -4,7 +4,7 @@ import java.io.{File, FileNotFoundException}
 
 import com.intellij.execution.process.{OSProcessHandler, ProcessAdapter, ProcessEvent}
 import com.intellij.openapi.util.Key
-import org.jetbrains.plugins.scala.compiler.HydraCredentialsManager
+import org.jetbrains.plugins.hydra.compiler.HydraCredentialsManager
 import org.jetbrains.plugins.scala.project.Platform
 
 /**
